adding test for filter function of dataset of local data class and other stuff
