+++ kotlin-spark-api/3.2/src/main/resources/META-INF/kotlin-jupyter-libraries/libraries.json
@@ -2,7 +2,7 @@
   "definitions": [],
   "producers": [
     {
-      "fqn": "org.jetbrains.kotlinx.spark.api.jupyter.Integration"
+      "fqn": "org.jetbrains.kotlinx.spark.api.jupyter.SparkIntegration"
     }
   ]
 }
