+++ pom_2.12.xml
@@ -14,6 +14,7 @@
   <packaging>pom</packaging>
 
   <properties>
+    <failOnMultipleScalaVersions>true</failOnMultipleScalaVersions>
     <scala.version>2.12.15</scala.version>
     <scala.compat.version>2.12</scala.compat.version>
   </properties>
